Add BusinessNetworkCardStore.getDefaultCardName static function
Common code for default naming of cards in a card store, for use by both the CLI
and Playground to ensure consistent default naming behaviour.
